requirements.txt
Adafruit-Blinka==8.19.0
adafruit-circuitpython-bno055==5.4.9
adafruit-circuitpython-busdevice==5.2.4
adafruit-circuitpython-register==1.9.15
adafruit-circuitpython-requests==1.13.2
adafruit-circuitpython-typing==1.9.2
Adafruit-PlatformDetect==3.45.1
Adafruit-PureIO==1.1.10
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==5.1.1
elevenlabs==0.2.10
executing==1.2.0
frozenlist==1.3.3
hcsr04sensor==1.6.3
hx711==1.1.2.3
idna==3.4
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
multidict==6.0.4
openai==0.27.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pigpio==1.78
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.13
pydantic==1.10.7
pydub==0.25.1
pyftdi==0.54.0
Pygments==2.15.1
pyserial==3.5
python-dotenv==1.0.0
pyusb==1.2.1
requests==2.29.0
RPi.GPIO==0.7.1
six==1.16.0
SpeechRecognition==3.10.0
stack-data==0.6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.9.2
